requirements/base.txt
Django==1.8.4
bpython==0.14.2
django-braces==1.8.0
django-model-utils==2.2
django-import-export==0.2.8
#logutils==0.3.3
#ipython==3.1.0
#psycopg2==2.6
django-bootstrap-form==3.1
Pillow==2.8.1
django-floppyforms==1.4.1
django-allauth==0.23.0
django-pure-pagination==0.2.1
